Fix: [Linux] Handle "Notify(Un)grab" for "FocusIn"-events
Fixes a problematic Event-emission "AppResume AppSuspend AppResume" when Alt-Tabbing-In to the app. This would not be visible to code just checking "AppSuspended()" as the final state is "suspended". But eventlisteners to "EVENT_APPRESUME" and "EVENT_APPSUSPEND" would get triggered (in this example this leads to flushing the keys on "app resume" too - as it emits the suspend-event).